Fix image uploading to S3
